push di prova

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\StoreCategoryRequest;
 use App\Http\Requests\UpdateCategoryRequest;
 use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
@@ -25,6 +26,10 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        $currentUserId = Auth::id();
+        if ($currentUserId != 1) {
+            abort(403);
+        }
         return view('admin.categories.create');
     }
 
@@ -53,6 +58,10 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
+        $currentUserId = Auth::id();
+        if ($currentUserId != 1) {
+            abort(403);
+        }
         return view('admin.categories.edit', compact('category'));
     }
 
@@ -76,6 +85,10 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        $currentUserId = Auth::id();
+        if ($currentUserId != 1) {
+            abort(403);
+        }
         $category->delete();
         return to_route('admin.categories.index')->with('message', "$category->title eliminato con successo!");
     }
